requirements/base.txt
Flask==3.0.0
gspread==5.12.0
oauth2client==4.1.3
pandas==2.1.4
plotly==5.17.0
python-dotenv==1.0.0
requests==2.31.0 